Make image list from UI match release page list
